+++ app/madmin/resources/company_resource.rb
@@ -9,7 +9,10 @@ class CompanyResource < Madmin::Resource
   attribute :updated_at, form: false
 
   # Associations
+  attribute :users
   attribute :teams
+  attribute :team_members
+  attribute :twitter_credentials
 
   # Add scopes to easily filter records
   # scope :published
@@ -20,10 +23,16 @@ class CompanyResource < Madmin::Resource
   # end
 
   # Customize the display name of records in the admin area.
-  # def self.display_name(record) = record.name
+  def self.display_name(record)
+    record.name
+  end
 
   # Customize the default sort column and direction.
-  # def self.default_sort_column = "created_at"
-  #
-  # def self.default_sort_direction = "desc"
+  def self.default_sort_column
+    "created_at"
+  end
+
+  def self.default_sort_direction
+    "desc"
+  end
 end