+++ app/madmin/resources/jwt_denylist_resource.rb
@@ -20,10 +20,16 @@ class JwtDenylistResource < Madmin::Resource
   # end
 
   # Customize the display name of records in the admin area.
-  # def self.display_name(record) = record.name
+  def self.display_name(record)
+    "JWT #{record.jti}"
+  end
 
   # Customize the default sort column and direction.
-  # def self.default_sort_column = "created_at"
-  #
-  # def self.default_sort_direction = "desc"
+  def self.default_sort_column
+    "created_at"
+  end
+
+  def self.default_sort_direction
+    "desc"
+  end
 end