+++ app/madmin/resources/user_resource.rb
@@ -5,14 +5,21 @@ class UserResource < Madmin::Resource
   # Attributes
   attribute :id, form: false
   attribute :email
+  attribute :username
+  attribute :did_tutorial
   attribute :remember_created_at
+  attribute :reset_password_sent_at
+  attribute :reset_password_token
   attribute :created_at, form: false
   attribute :updated_at, form: false
-  attribute :username
-  attribute :role
 
   # Associations
+  attribute :company
   attribute :team_member
+  attribute :team
+  attribute :strategies
+  attribute :posts
+  attribute :roles
 
   # Add scopes to easily filter records
   # scope :published
@@ -23,10 +30,16 @@ class UserResource < Madmin::Resource
   # end
 
   # Customize the display name of records in the admin area.
-  # def self.display_name(record) = record.name
+  def self.display_name(record)
+    record.username || record.email
+  end
 
   # Customize the default sort column and direction.
-  # def self.default_sort_column = "created_at"
-  #
-  # def self.default_sort_direction = "desc"
+  def self.default_sort_column
+    "created_at"
+  end
+
+  def self.default_sort_direction
+    "desc"
+  end
 end