+++ app/models/team.rb
@@ -21,4 +21,9 @@
 class Team < ApplicationRecord
   belongs_to :company
   has_many :team_members, dependent: :destroy
+  has_many :users, through: :team_members
+  has_many :strategies, through: :team_members
+  has_many :posts, through: :team_members
+
+  # Add validations as needed
 end