+++ db/migrate/20250616173546_add_did_tutorial_to_users.rb
@@ -1,0 +1,5 @@
+class AddDidTutorialToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :did_tutorial, :boolean, default: false, null: false
+  end
+end