+++ spec/controllers/api/v1/companies_controller_spec.rb
@@ -1,16 +1,18 @@
 require 'rails_helper'
 
 RSpec.describe Api::V1::CompaniesController do
-  let(:team) { create(:team) }
-  let(:team_employee) { create(:team) }
-  let(:admin) { create(:user, role: 'ADMIN', username: 'admin_user', email: 'admin@example.com') }
-  let(:employee) { create(:user, role: 'EMPLOYEE', username: 'emp_user', email: 'employee@example.com') }
-  let(:admin_team_member) { create(:team_member, user: admin, team: team) }
-  let(:employee_team_member) { create(:team_member, user: employee, team: team_employee) }
   let!(:company) { create(:company) }
   let!(:company_employee) { create(:company) }
-  let!(:admin_company_member) { create(:company_member, user: admin, company: company) }
-  let!(:employee_company_member) { create(:company_member, user: employee, company: company_employee) }
+
+  let(:admin) { create(:user, role: 'ADMIN', username: 'admin_user', email: 'admin@example.com') }
+  let(:employee) { create(:user, role: 'EMPLOYEE', username: 'emp_user', email: 'employee@example.com') }
+
+  let!(:admin_team) { create(:team, company: company) }
+  let!(:employee_team) { create(:team, company: company_employee) }
+
+  let!(:admin_team_member) { create(:team_member, user: admin, team: admin_team) }
+  let!(:employee_team_member) { create(:team_member, user: employee, team: employee_team) }
+
   let!(:twitter_credential) { create(:twitter_credential, company: company) }
   let!(:twitter_credential_employee) { create(:twitter_credential, company: company_employee) }
 
@@ -22,7 +24,7 @@ RSpec.describe Api::V1::CompaniesController do
       end
 
       it 'returns a successful response with social network status' do
-        get :company_social_status
+        get :social_network_status
         expect(response).to have_http_status(:ok)
         json_response = JSON.parse(response.body)
         expect(json_response['status']['code']).to eq(200)
@@ -33,7 +35,7 @@ RSpec.describe Api::V1::CompaniesController do
 
     context 'when the user is not authenticated' do
       it 'returns an unauthorized response' do
-        get :company_social_status
+        get :social_network_status
         expect(response).to have_http_status(:unauthorized)
       end
     end
@@ -47,11 +49,11 @@ RSpec.describe Api::V1::CompaniesController do
       end
 
       it 'returns a not found response' do
-        get :company_social_status
+        get :social_network_status
         expect(response).to have_http_status(:not_found)
         json_response = JSON.parse(response.body)
         expect(json_response['status']['code']).to eq(422)
-        expect(json_response['errors']).to include('User is not associated with any company')
+        expect(json_response['errors']).to include('User is not associated with a company.')
       end
     end
   end
@@ -84,7 +86,7 @@ RSpec.describe Api::V1::CompaniesController do
         expect(response).to have_http_status(:not_found)
         json_response = JSON.parse(response.body)
         expect(json_response['status']['code']).to eq(422)
-        expect(json_response['errors']).to include('Company not found')
+        expect(json_response['errors']).to include('Company not found.')
       end
     end